+++ tests/headers/test_ressource_provider.hpp
@@ -24,11 +24,17 @@
 
 #include <gtest/gtest.h>
 
+#include <utility/ressource_provider.hpp>
+#include <utility/system_io/default_system_io.hpp>
+
 namespace tests::utility
 {
 	class TestRessourceProvider: public ::testing::Test
 	{
 		protected:
+		::utility::DefaultSystemIO _systemIO;
+		::utility::RessourceProvider _provider { _systemIO };
+
 		void SetUp(void) override
 		{
 		}